Service extraction and error truncate
* Moved out coveralls service call
* added truncate to service error response
