+++ src/csmacnz.Coveralls/ICoverallsService.cs
@@ -1,0 +1,9 @@
+using Beefeater;
+
+namespace csmacnz.Coveralls
+{
+    public interface ICoverallsService
+    {
+        Result<bool, string> Upload(string fileData);
+    }
+}
